[clangd] Add AST config to prebuild ASTs
The new AST config allows prebuilding the ASTs for files in a
compilation database when a single file in that compilation database
is opened by the client. This reduces latency when opening subsequent
files in the same compilation database. This matters in large
codebases where building the AST can take multiple seconds.